It's been far too long since I've checked anything in. This is the version I am currently working with. Note that the information node's are not correct anymore as I accidentally wiped the art I modified to make them work. If I can get the art that has no white space around it as a spritesheet, that would be handy! :)

+++ trunk/Source/Fog.cpp
@@ -49,7 +49,7 @@ void Fog::initFog(void)
 	{
 		for (int j = 0; j < 5; j++)
 		{
-			Point Location(i, j);
+			Point Location((float) i, (float) j);
 
 			FogBank[i][j] = AnimatedSprite(64, 128, 11, FogSprite.getSpriteNumber());
 			FogBank[i][j].setDepth(10);
